requirements.txt
langchain
chromadb
openai
pymupdf
streamlit
python-dotenv
langchain_community
langchain_openai